move color conversion to separate file
